# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GeoTandem",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/layers/{name}": {
      "get": {
        "summary": "Layer",
        "operationId": "layer_api_layers__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayerInfo"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/api/query": {
      "post": {
        "summary": "Query",
        "description": "Run a query object (F-8.9); the result carries its query and provenance.",
        "operationId": "query_api_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryObject-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/query/count": {
      "post": {
        "summary": "Count",
        "description": "Count the features of several queries at once, e.g. one per condition (design B2).\n\nNothing but numbers leaves the server. A rejected query names its position\nin ``details.index``. All of them share one time limit, as one query would:\na request of many slow counts must not hold a worker many times as long.",
        "operationId": "count_api_query_count_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Counts"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/query/ids": {
      "post": {
        "summary": "Ids",
        "description": "The ids of the features a query returns, for marking hits (design B9).\n\nNo geometry or attribute is sent, so the result-size limit does not apply.",
        "operationId": "ids_api_query_ids_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryObject-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ids"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/query/validate": {
      "post": {
        "summary": "Validate",
        "description": "Check a query against schema and data core without running it (F-5.9).",
        "operationId": "validate_api_query_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryObject-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Validation"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttributeInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "data_type": {
            "type": "string",
            "enum": [
              "integer",
              "real",
              "text",
              "boolean",
              "date"
            ],
            "title": "Data Type"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value_domain": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Domain"
          },
          "for_model": {
            "type": "boolean",
            "title": "For Model"
          },
          "references": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "References"
          }
        },
        "type": "object",
        "required": [
          "name",
          "data_type",
          "label",
          "description",
          "unit",
          "value_domain",
          "for_model",
          "references"
        ],
        "title": "AttributeInfo"
      },
      "AttributeJoin": {
        "properties": {
          "layer": {
            "type": "string",
            "pattern": "^[a-z_][a-z0-9_]{0,62}$",
            "title": "Layer",
            "description": "Lower-case layer or attribute name."
          },
          "left_key": {
            "type": "string",
            "pattern": "^[a-z_][a-z0-9_]{0,62}$",
            "title": "Left Key",
            "description": "Attribute of the source layer."
          },
          "right_key": {
            "type": "string",
            "pattern": "^[a-z_][a-z0-9_]{0,62}$",
            "title": "Right Key",
            "description": "Attribute of the joined layer."
          },
          "fields": {
            "items": {
              "type": "string",
              "pattern": "^[a-z_][a-z0-9_]{0,62}$",
              "description": "Lower-case layer or attribute name."
            },
            "type": "array",
            "minItems": 1,
            "title": "Fields"
          },
          "prefix": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-z_][a-z0-9_]{0,30}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefix"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "layer",
          "left_key",
          "right_key",
          "fields"
        ],
        "title": "AttributeJoin",
        "description": "Left join of a layer by key (F-4.6).\n\nJoined fields are addressable by their name, or by ``prefix + name`` when a\nprefix is given; clashes with source attributes are rejected."
      },
...
      "Between": {
        "properties": {
          "op": {
            "type": "string",
            "const": "between",
            "title": "Op"
          },
          "attr": {
            "type": "string",
            "pattern": "^[a-z_][a-z0-9_]{0,62}$",
            "title": "Attr",
            "description": "Lower-case layer or attribute name."
          },
          "min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              }
            ],
            "title": "Min"
          },
          "max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              }
            ],
            "title": "Max"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "op",
          "attr",
          "min",
          "max"
        ],
        "title": "Between",
        "description": "Inclusive value range (F-4.2). The bounds may come in either order."
      },
...
      "Compare": {
        "properties": {
          "op": {
            "type": "string",
            "const": "compare",
            "title": "Op"
          },
          "attr": {
            "type": "string",
            "pattern": "^[a-z_][a-z0-9_]{0,62}$",
            "title": "Attr",
            "description": "Lower-case layer or attribute name."
          },
          "cmp": {
            "type": "string",
            "enum": [
              "eq",
              "ne",
              "lt",
              "le",
              "gt",
              "ge"
            ],
            "title": "Cmp"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              }
            ],
            "title": "Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "op",
          "attr",
          "cmp",
          "value"
        ],
        "title": "Compare",
        "description": "Attribute comparison with a constant (F-4.2)."
      },
...
      "CountRequest": {
        "properties": {
          "queries": {
            "items": {
              "$ref": "#/components/schemas/QueryObject-Input"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Queries"
          }
        },
        "type": "object",
        "required": [
          "queries"
        ],
        "title": "CountRequest"
      },
      "Counts": {
        "properties": {
          "counts": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Counts"
          }
        },
        "type": "object",
        "required": [
          "counts"
        ],
        "title": "Counts"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
...
      "Ids": {
        "properties": {
          "ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Ids"
          }
        },
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "Ids"
      },
...
      "LayerInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "kind": {
            "type": "string",
            "enum": [
              "vector",
              "table"
            ],
            "title": "Kind"
          },
          "geometry_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry Type"
          },
          "feature_count": {
            "type": "integer",
            "title": "Feature Count"
          },
          "bbox_wgs84": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox Wgs84"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "dataset_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Version"
          },
          "for_model": {
            "type": "boolean",
            "title": "For Model"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "attributes": {
            "items": {
              "$ref": "#/components/schemas/AttributeInfo"
            },
            "type": "array",
            "title": "Attributes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "title",
          "description",
          "kind",
          "geometry_type",
          "feature_count",
          "bbox_wgs84",
          "source",
          "dataset_version",
          "for_model",
          "created_at",
          "updated_at",
          "attributes"
        ],
        "title": "LayerInfo"
      },
...
      "QueryObject-Input": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "2",
            "title": "Schema Version",
            "default": "2"
          },
          "source": {
            "type": "string",
            "pattern": "^[a-z_][a-z0-9_]{0,62}$",
            "title": "Source",
            "description": "Lower-case layer or attribute name."
          },
          "attribute_join": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AttributeJoin"
              },
              {
                "type": "null"
              }
            ]
          },
          "where": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Compare"
                  },
                  {
                    "$ref": "#/components/schemas/Between"
                  },
                  {
                    "$ref": "#/components/schemas/InList"
                  },
                  {
                    "$ref": "#/components/schemas/TextMatch"
                  },
                  {
                    "$ref": "#/components/schemas/IsNull"
                  },
                  {
                    "$ref": "#/components/schemas/BBox"
                  },
                  {
                    "$ref": "#/components/schemas/GeometryFilter"
                  },
                  {
                    "$ref": "#/components/schemas/NearFeature"
                  },
                  {
                    "$ref": "#/components/schemas/RelatedTopological-Input"
                  },
                  {
                    "$ref": "#/components/schemas/RelatedByDistance-Input"
                  },
                  {
                    "$ref": "#/components/schemas/And-Input"
                  },
                  {
                    "$ref": "#/components/schemas/Or-Input"
                  },
                  {
                    "$ref": "#/components/schemas/Not-Input"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Where"
          },
          "buffer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Buffer"
              },
              {
                "type": "null"
              }
            ]
          },
          "spatial_relation": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/TopologicalRelation-Input"
                  },
                  {
                    "$ref": "#/components/schemas/DistanceRelation-Input"
                  }
                ],
                "discriminator": {
                  "propertyName": "predicate",
                  "mapping": {
                    "contains": "#/components/schemas/TopologicalRelation-Input",
                    "dwithin": "#/components/schemas/DistanceRelation-Input",
                    "intersects": "#/components/schemas/TopologicalRelation-Input",
                    "within": "#/components/schemas/TopologicalRelation-Input"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Spatial Relation"
          },
          "aggregate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Aggregate"
              },
              {
                "type": "null"
              }
            ]
          },
          "columns": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/DistanceColumn-Input"
                },
                {
                  "$ref": "#/components/schemas/ValueColumn-Input"
                }
              ],
              "discriminator": {
                "propertyName": "fn",
                "mapping": {
                  "distance_to": "#/components/schemas/DistanceColumn-Input",
                  "value_of": "#/components/schemas/ValueColumn-Input"
                }
              }
            },
            "type": "array",
            "title": "Columns"
          },
          "select": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "pattern": "^[a-z_][a-z0-9_]{0,62}$",
                  "description": "Lower-case layer or attribute name."
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Select",
            "description": "Result attributes; all when omitted."
          },
          "order_by": {
            "items": {
              "$ref": "#/components/schemas/OrderBy"
            },
            "type": "array",
            "title": "Order By"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "symbology": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/SingleSymbol"
                  },
                  {
                    "$ref": "#/components/schemas/Categorized"
                  },
                  {
                    "$ref": "#/components/schemas/Classified"
                  },
                  {
                    "$ref": "#/components/schemas/GraduatedSize"
                  }
                ],
                "discriminator": {
                  "propertyName": "kind",
                  "mapping": {
                    "categorized": "#/components/schemas/Categorized",
                    "classified": "#/components/schemas/Classified",
                    "graduated_size": "#/components/schemas/GraduatedSize",
                    "single": "#/components/schemas/SingleSymbol"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbology"
          },
          "output": {
            "type": "string",
            "enum": [
              "map",
              "table"
            ],
            "title": "Output",
            "default": "map"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source"
        ],
        "title": "GeoTandem query object",
        "description": "Declarative, backend-neutral description of one analysis."
      },
...
      "QueryResult": {
        "properties": {
          "type": {
            "type": "string",
            "const": "FeatureCollection",
            "title": "Type",
            "default": "FeatureCollection"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/Feature"
            },
            "type": "array",
            "title": "Features"
          },
          "query": {
            "$ref": "#/components/schemas/QueryObject-Output"
          },
          "meta": {
            "$ref": "#/components/schemas/ResultMeta"
          }
        },
        "type": "object",
        "required": [
          "features",
          "query",
          "meta"
        ],
        "title": "QueryResult",
        "description": "A GeoJSON FeatureCollection plus the query and its provenance."
      },
...
      "Validation": {
        "properties": {
          "valid": {
            "type": "boolean",
            "const": true,
            "title": "Valid",
            "default": true
          },
          "layers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Layers"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/Op"
            },
            "type": "array",
            "title": "Operations"
          }
        },
        "type": "object",
        "required": [
          "layers",
          "operations"
        ],
        "title": "Validation"
      },
...
    }
  }
}